Version 0x11 released: Instead of shutting off RC servo outputs when safety timeout expires, we now 'center' them at 1.5ms pulses (0x80)

+++ firmware/src/RCMx1.X/user.c
@@ -45,8 +45,8 @@
 /* Global variable storage                                                    */
 /******************************************************************************/
 
-// Safety Timeout Value reigster
-UINT8 SafetyTimeoutValue = 0;
+// Safety Timeout Value register
+UINT16 SafetyTimeoutValue = 0;
 
 // GPIO pins 1-6
 UINT8 GPIO_IO = 0;
